Removed entries from zod custom json. Refactor auth lib dir structure
