+++ src/lib/authentication/unauthorized.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Box, Heading, Text, Button } from "@chakra-ui/react";
-import { useAuthContext } from "../../components/useAuthContext";
-import { useAppDispatch } from "../../../../app/hooks";
+import { useAuthContext } from "./components/useAuthContext";
+import { useAppDispatch } from "../../app/hooks";
 import { Login } from "./authenticationSlice";
 
 const Unauthorized = () => {
